+++ LuckyoneUI/Profiles/ElvUI.lua
@@ -314,6 +314,8 @@ function Private:Setup_NamePlates(installer)
 	E.db.nameplates.overlapV = 1.6
 	E.db.nameplates.statusbar = Private.Texture
 	E.db.nameplates.threat.useSoloColor = true
+	E.db.nameplates.threat.skipGoodColor = true
+	E.db.nameplates.useBlizzardAuras = Private.isRetail -- Use filter prio in nonRetail
 
 	-- NamePlates misc
 	E.db.nameplates.classColorNames = true